src: More bugfixes, nearly finished migration script

+++ src/lib/types/settings/raw/RawMemberSettings.ts
@@ -8,7 +8,7 @@ export const SQL_TABLE_SCHEMA = /* sql */`
 	CREATE TABLE IF NOT EXISTS members (
 		"guild_id"    VARCHAR(19)           NOT NULL,
 		"user_id"     VARCHAR(19)           NOT NULL,
-		"point_count" INTEGER     DEFAULT 0 NOT NULL CHECK(point_count >= 0),
+		"point_count" BIGINT      DEFAULT 0 NOT NULL CHECK(point_count >= 0),
 		CONSTRAINT members_guild_user_idx   PRIMARY KEY("guild_id", "user_id")
 	);
